Rename vpu_patterns to npu_patterns in tools/pot
