+++ tools/pot/openvino/tools/pot/graph/npu_patterns.py
