+++ tools/pot/openvino/tools/pot/graph/utils.py
@@ -10,7 +10,7 @@ import numpy as np
 from openvino.tools.pot.version import get_version
 from .cpu_patterns import get_cpu_ignored_patterns, get_cpu_spr_ignored_patterns
 from .gpu_patterns import get_gpu_ignored_patterns
-from .vpu_patterns import get_npu_ignored_patterns
+from .npu_patterns import get_npu_ignored_patterns
 from .gna_patterns import get_gna_ignored_patterns, get_gna3_ignored_patterns
 from .special_operations import QUANTIZE_AGNOSTIC_OPERATIONS
 from .node_utils import get_all_node_outputs, get_input_shape
